Add ESX variant using Betweeness-Centrality heuristic
Now ESX can use edges Betweeness-Centrality as heuristic to select the
next edge to hide from search graph. This should result in a dramatic
speedup for ESX. Unit tests are provided also.

+++ test/test_types.hpp
@@ -5,10 +5,9 @@
 
 namespace arlib {
 namespace test {
-using Graph =
-    boost::adjacency_list<boost::vecS, boost::vecS, boost::bidirectionalS,
-                          boost::no_property,
-                          boost::property<boost::edge_weight_t, double>>;
+using Graph = boost::adjacency_list<boost::vecS, boost::vecS,
+                                    boost::bidirectionalS, boost::no_property,
+                                    boost::property<boost::edge_weight_t, int>>;
 using Vertex = boost::graph_traits<Graph>::vertex_descriptor;
 using Edge = boost::graph_traits<Graph>::edge_descriptor;
 using Length = typename boost::property_traits<typename boost::property_map<
